Create a dedicated module for RulerToolButton
